+++ rosbot_controller/src/mppi/losses/losses.py
@@ -2,6 +2,7 @@ import numpy as np
 from numba import jit
 from scipy.spatial import distance
 
+import sys
 
 def sum_loss(trajectories, reference_traj, traj_lookahead, goal_idx):
     loss = np.zeros(shape=(trajectories.shape[0], trajectories.shape[1]))
