+++ rosbot_controller/src/mppi/mppic.py
@@ -14,11 +14,11 @@ from geometry_msgs.msg import Vector3, Point
 from nav_msgs.msg import Path
 from visualization_msgs.msg import Marker, MarkerArray
 
-from mpc_dtypes import State, Control, dist_L2, dist_L2_np
-from mpc_utils import quaternion_to_euler
-from models import Rosbot
+from utils.mpc_dtypes import State, Control, dist_L2, dist_L2_np
+from utils.mpc_utils import quaternion_to_euler
+from models.models import Rosbot
 
-from losses import sum_loss, order_loss, nearest_loss
+from losses.losses import sum_loss, order_loss, nearest_loss
 
 
 class MPPIController:
